pyproject.toml
[project]
name = "gptchat"
version = "0.1.0"
description = ""
authors = [
    {name = "gollum",email = "gollum80@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai (>=1.77.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "gptchat"}]


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
